Make license more clear
I wish it was more clear the correct way to state multiple licenses for different
parts of a repository. I don't see licensing every being an issue, and don't plan
on going after license-breakers, or expect anyone to come after me. Hopefully
this is the last time I worry about it.